+++ sql/sql_repl.cc
@@ -3782,7 +3782,7 @@ bool change_master(THD* thd, Master_info* mi, bool *master_info_added)
   if (lex_mi->connect_retry)
     mi->connect_retry = lex_mi->connect_retry;
   if (lex_mi->heartbeat_opt != LEX_MASTER_INFO::LEX_MI_UNCHANGED)
-    mi->heartbeat_period = lex_mi->heartbeat_period;
+    mi->heartbeat_period = static_cast<uint32_t>(lex_mi->heartbeat_period);
   else
     mi->heartbeat_period= MY_MIN(SLAVE_MAX_HEARTBEAT_PERIOD,
                                  slave_net_timeout*500ULL);
